+++ types/locations/location-bar.d.ts
@@ -1,5 +1,4 @@
-export default History;
-declare class History {
+export class History {
     handlers: any[];
     checkUrl(): boolean;
     location: Location;
@@ -7,14 +6,10 @@ declare class History {
     atRoot(): boolean;
     getHash(): string;
     getFragment(fragment: any, forcePushState: any): any;
-    start(options?: {}): boolean;
+    start(options: any): boolean;
     started: boolean;
-    options: {
-        root: string;
-    };
     root: any;
     _wantsHashChange: boolean;
-    _wantsPushState: boolean;
     _hasPushState: boolean;
     fragment: any;
     stop(): void;
@@ -23,6 +18,5 @@ declare class History {
     update(fragment: any, options: any): boolean | void;
     _updateHash(location: any, fragment: any, replace: any): void;
     onChange(callback: any): void;
-    hasPushState(): boolean;
 }
 //# sourceMappingURL=location-bar.d.ts.map