+++ types/locations/memory.d.ts
@@ -1,16 +1,14 @@
 export default MemoryLocation;
 declare class MemoryLocation {
-    constructor({ path }: {
-        path: any;
-    });
-    path: any;
-    getURL(): any;
+    path: string;
+    getURL(): string;
     setURL(path: any, options: any): void;
     replaceURL(path: any, options: any): void;
     onChange(callback: any): void;
-    changeCallback: any;
-    handleURL(url: any, options?: {}): void;
+    callback: any;
+    handleURL(url: any, options: any): void;
     removeRoot(url: any): any;
     formatURL(url: any): any;
+    start(path: any): void;
 }
 //# sourceMappingURL=memory.d.ts.map