+++ types/transition.d.ts
@@ -5,7 +5,7 @@
  */
 export default function transition(options: any): Transition;
 export type Route = import("./router.js").Route;
-export type TransitionData = Pick<Transition, 'routes' | 'pathname' | 'path' | 'params' | 'query'>;
+export type TransitionData = Pick<Transition, "routes" | "pathname" | "path" | "params" | "query">;
 export type Transition = {
     routes: Route[];
     pathname: string;